pyproject.toml
[tool.poetry]
name = "spidey-kafka"
version = "0.1.0"
description = ""
authors = ["Arturo Cuicas <orutrax@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.5"
python-dotenv = "^1.0.0"
aiokafka = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
